backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

pandas==2.2.2
numpy==1.26.4
networkx==3.3

scikit-learn==1.5.0
joblib==1.4.2

sqlalchemy==2.0.30
pydantic==2.7.4
python-multipart==0.0.9
python-dotenv==1.0.1

typing-extensions==4.12.2
